# doc/api/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    AttachmentResponse:
      type: object
      required:
      - id
      - fileName
      - contentType
      - sizeBytes
      properties:
        id:
          type: integer
          format: int64
          examples:
          - 7
        fileName:
          type: string
          examples:
          - Rechnung.pdf
          maxLength: 255
        contentType:
          type: string
          examples:
          - application/pdf
          maxLength: 255
        sizeBytes:
          type: integer
          format: int64
          examples:
          - 48213
...
    LocalDate:
      type: string
      format: date
...
    TaskCreateRequest:
      type: object
      required:
      - description
      - dueDate
      - importance
      - state
      properties:
        description:
          type: string
          examples:
          - Renew the passport
          pattern: \S
          maxLength: 255
        dueDate:
          type: string
          examples:
          - 2026-11-30
          $ref: "#/components/schemas/LocalDate"
        importance:
          type: string
          examples:
          - HIGH
          $ref: "#/components/schemas/TaskImportance"
        state:
          type: string
          examples:
          - TODO
          $ref: "#/components/schemas/TaskState"
        attachmentIds:
          type:
          - array
          - "null"
          items:
            type: integer
            format: int64
          description: "Attachments of a task deleted moments ago, for an undo to\
            \ bring back (#204); only the caller's, only within the undo window."
          maxItems: 10
    TaskImportance:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
    TaskResponse:
      type: object
      required:
      - id
      - description
      - dueDate
      - importance
      - state
      properties:
        id:
          type: integer
          format: int64
          examples:
          - 42
        description:
          type: string
          examples:
          - Renew the passport
          maxLength: 255
        dueDate:
          type: string
          examples:
          - 2026-11-30
          $ref: "#/components/schemas/LocalDate"
        importance:
          type: string
          examples:
          - HIGH
          $ref: "#/components/schemas/TaskImportance"
        state:
          type: string
          examples:
          - TODO
          $ref: "#/components/schemas/TaskState"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentResponse"
          description: Its available attachments (#204). Always sent; optional in
            the schema so the board still validates a task from a backend rolled back
            to before attachments.
    TaskState:
      type: string
      enum:
      - TODO
      - WORKING
      - DONE
    TaskUpdateRequest:
      type: object
      required:
      - description
      - dueDate
      - importance
      - state
      properties:
        description:
          type: string
          examples:
          - Renew the passport
          pattern: \S
          maxLength: 255
        dueDate:
          type: string
          examples:
          - 2026-11-30
          $ref: "#/components/schemas/LocalDate"
        importance:
          type: string
          examples:
          - HIGH
          $ref: "#/components/schemas/TaskImportance"
        state:
          type: string
          examples:
          - TODO
          $ref: "#/components/schemas/TaskState"
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
      description: Authentication
...
paths:
...
  /api/tasks:
    get:
      summary: List the caller's tasks
      description: "Ordered by due date, then id."
      tags:
      - Tasks
      responses:
        "200":
          description: The caller's tasks.
          content:
            application/json:
              examples:
                tasks:
                  value:
                  - id: 42
                    description: Renew the passport
                    dueDate: 2026-11-30
                    importance: HIGH
                    state: TODO
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
    post:
      summary: Create a task
      description: The caller becomes the owner; the id cannot be set.
      tags:
      - Tasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
        required: true
      responses:
        "201":
          description: The task as created.
          content:
            application/json:
              examples:
                task:
                  value:
                    id: 42
                    description: Renew the passport
                    dueDate: 2026-11-30
                    importance: HIGH
                    state: TODO
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          description: "A field failed validation, for example a blank description\
            \ or a due date in the past."
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/tasks/{id}:
    put:
      summary: Replace a task
      description: Replaces every field of a task owned by the caller.
      tags:
      - Tasks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
        required: true
      responses:
        "200":
          description: The task as updated.
          content:
            application/json:
              examples:
                task:
                  value:
                    id: 42
                    description: Renew the passport
                    dueDate: 2026-11-30
                    importance: HIGH
                    state: TODO
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          description: A field failed validation.
        "404":
          description: No task with this id is owned by the caller.
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
    delete:
      summary: Delete a task
      tags:
      - Tasks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: The task was deleted.
        "404":
          description: No task with this id is owned by the caller.
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
...
info:
  title: TaskFest API
  version: 1.0.0-SNAPSHOT
  description: The task board's REST surface. Every response schema here is published
    as JSON Schema and checked at runtime by the frontend.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
- url: /